+++ src/solverwrapper.cc
@@ -28,7 +28,7 @@ int construct_sol(Scheduleset **set, int *nnewsets, int *d, Optimal_Solution<T> 
         std::copy(v->begin(), v->end(), newset->members);
     }
 
-    for (int i = 0; i < sol.jobs.size(); i++) {
+    for (size_t i = 0; i < sol.jobs.size(); i++) {
         C += d[newset->members[i]];
         newset->C[i] = C;
         g_hash_table_insert(newset->table, GINT_TO_POINTER(newset->members[i]), newset->C + i);
@@ -264,25 +264,25 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                solver->init_bdd_table();
-                break;
+        case bdd_solver:
+            solver->init_bdd_table();
+            break;
 
-            case zdd_solver:
-                solver->init_zdd_table();
-                break;
+        case zdd_solver:
+            solver->init_zdd_table();
+            break;
 
-            case DP_solver:
-                break;
+        case DP_solver:
+            break;
         }
 
         switch (parms->construct) {
-            case yes_construct:
-                break;
+        case yes_construct:
+            break;
 
-            case no_construct:
-                solver->init_table_farkas();
-                break;
+        case no_construct:
+            solver->init_table_farkas();
+            break;
         }
 
         return val = 0;
@@ -293,16 +293,16 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                solver->init_bdd_conflict_solver(elist_same, ecount_same, elist_differ, ecount_differ);
-                break;
+        case bdd_solver:
+            solver->init_bdd_conflict_solver(elist_same, ecount_same, elist_differ, ecount_differ);
+            break;
 
-            case zdd_solver:
-                solver->init_zdd_conflict_solver(elist_same, ecount_same, elist_differ, ecount_differ);
-                break;
+        case zdd_solver:
+            solver->init_zdd_conflict_solver(elist_same, ecount_same, elist_differ, ecount_differ);
+            break;
 
-            case DP_solver:
-                break;
+        case DP_solver:
+            break;
         }
 
         return val;
@@ -323,16 +323,16 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                solver->free_bdd_solver(ecount_same, ecount_differ);
-                break;
+        case bdd_solver:
+            solver->free_bdd_solver(ecount_same, ecount_differ);
+            break;
 
-            case zdd_solver:
-                solver->free_zdd_solver(ecount_same, ecount_differ);
-                break;
+        case zdd_solver:
+            solver->free_zdd_solver(ecount_same, ecount_differ);
+            break;
 
-            case DP_solver:
-                break;
+        case DP_solver:
+            break;
         }
 
         return val;
@@ -358,16 +358,16 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                solver->init_bdd_one_conflict(v1, v2, same);
-                break;
+        case bdd_solver:
+            solver->init_bdd_one_conflict(v1, v2, same);
+            break;
 
-            case zdd_solver:
-                solver->init_zdd_one_conflict(v1, v2, same);
-                break;
+        case zdd_solver:
+            solver->init_zdd_one_conflict(v1, v2, same);
+            break;
 
-            case DP_solver:
-                break;
+        case DP_solver:
+            break;
         }
 
         return val;
@@ -427,20 +427,20 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                val = solve_weight_dbl_bdd(pd);
-                CCcheck_val_2(val, "Failed solve_weight_dbl_bdd");
-                break;
+        case bdd_solver:
+            val = solve_weight_dbl_bdd(pd);
+            CCcheck_val_2(val, "Failed solve_weight_dbl_bdd");
+            break;
 
-            case zdd_solver:
-                val = solve_weight_dbl_zdd(pd);
-                CCcheck_val_2(val, "Failed solve_weight_dbl_zdd")
-                break;
+        case zdd_solver:
+            val = solve_weight_dbl_zdd(pd);
+            CCcheck_val_2(val, "Failed solve_weight_dbl_zdd")
+            break;
 
-            case DP_solver:
-                val = solve_dynamic_programming_ahv(pd);
-                CCcheck_val_2(val, "Failed in solve_dynamic_programming_ahv")
-                break;
+        case DP_solver:
+            val = solve_dynamic_programming_ahv(pd);
+            CCcheck_val_2(val, "Failed in solve_dynamic_programming_ahv")
+            break;
         }
 
 CLEAN:
@@ -452,20 +452,20 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-                val = solve_weight_dbl_bdd_CG_heur(data);
-                CCcheck_val_2(val, "Failed solve_weight_dbl_bdd");
-                break;
+        case bdd_solver:
+            val = solve_weight_dbl_bdd_CG_heur(data);
+            CCcheck_val_2(val, "Failed solve_weight_dbl_bdd");
+            break;
 
-            case zdd_solver:
-                val = solve_weight_dbl_zdd_CG_heur(data);
-                CCcheck_val_2(val, "Failed solve_weight_dbl_zdd")
-                break;
+        case zdd_solver:
+            val = solve_weight_dbl_zdd_CG_heur(data);
+            CCcheck_val_2(val, "Failed solve_weight_dbl_zdd")
+            break;
 
-            case DP_solver:
-                val = solve_dynamic_programming_ahv_CG_heur(data);
-                CCcheck_val_2(val, "Failed in solve_dynamic_programming_ahv")
-                break;
+        case DP_solver:
+            val = solve_dynamic_programming_ahv_CG_heur(data);
+            CCcheck_val_2(val, "Failed in solve_dynamic_programming_ahv")
+            break;
         }
 
 CLEAN:
@@ -540,17 +540,17 @@ CLEAN:
     void compute_sol_stab(PricerSolver *solver, wctparms *parms, double *pi, Optimal_Solution<double> *sol)
     {
         switch (parms->solver) {
-            case bdd_solver:
-                *sol = solver->solve_weight_bdd_double(pi);
-                break;
+        case bdd_solver:
+            *sol = solver->solve_weight_bdd_double(pi);
+            break;
 
-            case zdd_solver:
-                *sol = solver->solve_weight_zdd_double(pi);
-                break;
+        case zdd_solver:
+            *sol = solver->solve_weight_zdd_double(pi);
+            break;
 
-            case DP_solver:
-                *sol = solver->dynamic_programming_ahv(pi);
-                break;
+        case DP_solver:
+            *sol = solver->dynamic_programming_ahv(pi);
+            break;
         }
     }
 
@@ -559,16 +559,16 @@ CLEAN:
         int val = 0;
 
         switch (parms->solver) {
-            case bdd_solver:
-            case zdd_solver:
-                val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->duration, sol, pd->njobs);
-                CCcheck_val_2(val, "Failed in construction of solution");
-                break;
+        case bdd_solver:
+        case zdd_solver:
+            val = construct_sol(&(pd->newsets), &(pd->nnewsets), pd->duration, sol, pd->njobs);
+            CCcheck_val_2(val, "Failed in construction of solution");
+            break;
 
-            case DP_solver:
-                val = construct_sol<double, true>(&(pd->newsets), &(pd->nnewsets), pd->duration, sol, pd->njobs);
-                CCcheck_val_2(val, "Failed in construction of solution");
-                break;
+        case DP_solver:
+            val = construct_sol<double, true>(&(pd->newsets), &(pd->nnewsets), pd->duration, sol, pd->njobs);
+            CCcheck_val_2(val, "Failed in construction of solution");
+            break;
         }
 
 CLEAN:
